Move SetSavedCharacterKnownFlag to bmsave.c
